[22] Improve the extensibility of the server
The server has been separated in multiple bundles. The original one is
now limited to the core behavior of the server. A Sirius specific bundle
has been introduced with servlets dedicated to Sirius and the Workflow
DSL. It may be necessary in the future to separate both parts but it
won't have any impact on the extension solution.

A bundle has been created with the business layer of the Sirius specific
servlets. Finally another bundle has been added to contain the exposed
API of the business layer. It will allow us to replace the real Sirius
services easily in order to test the server with fake ones.

The build has been improved because it is now depending on some Google
Guice fragments (multibindings and servlet) that have to referenced
directly from the pom.xml files.

Finally the code of the Sirius services has been improved in order to
leverage additional features of Java 8 optionals and streams.

Bug: https://github.com/ObeoNetwork/Jarvis/issues/22
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@gmail.com>